refactor(vk): move ReShade runtime state into swapchain
