+++ tests/meson.build
@@ -39,8 +39,8 @@ event_bus_tests = executable('event_bus_tests',
   include_directories : include_directories('../src/core')
 )
 
-reshade_runtime_tests = executable('reshade_runtime_tests',
-  'reshade_runtime_tests.cpp',
+reshade_uniform_tests = executable('reshade_uniform_tests',
+  'reshade_uniform_tests.cpp',
   dependencies : [
     catch2,
   ],
@@ -51,4 +51,4 @@ test('ConfigObserver', config_observer_tests)
 test('ConfigParser', config_parser_tests)
 test('ConfigStore', config_store_tests)
 test('EventBus', event_bus_tests)
-test('ReShadeRuntime', reshade_runtime_tests)
+test('ReshadeUniforms', reshade_uniform_tests)
